Changing data types, adding transformation of hex to uint8_t

+++ lymui/rgb.c
@@ -15,7 +15,7 @@
 // Make RGB
 // Return RGB struct pointer
 // Params : Array of ushort length 3
-struct Rgb *makeRGB(ushort *c) {
+struct Rgb *makeRGB(uint8_t *c) {
     struct Rgb *color = malloc (sizeof (struct Rgb));
     
     if (color == NULL)
